refactor: map wave-6 Slice C densify tests and 300-NCL ratchet (#4371)
Extract Pass 5 border noise, split oversized test/support files to ≤300
physical lines, and lower repo.map_lib_file_size_headroom 350→300.

Refs #4371
