+++ SPEC/program/repo-lint.md
@@ -81,7 +81,7 @@
 | `tool/check_map_gen_stage_protocol.dart` | Require the shared `MapGenStage` contract in `packages/colonizethis_map/lib/src/gen/map_gen_stage.dart` and `implements MapGenStage` (or the uniform `MapGenPass`) on each bound generator service family (`TileMapGenLandSeeds`, `TileMapGenLakesProvinces`, `ContinentJoinPass`, `TerrainJitterPass`, `SeaZoneSubdividePass`, `TileMapGenTerrainResource`) so pass orchestration shares a uniform params + grid-in/grid-out protocol, and require **at least 4** bound families to adopt the uniform `MapGenPass` entry point (raised from 3 after the former join-sea family was split into three standalone `MapGenPass` passes, removing the last `MapGenStage`-only exemption) (Refs #3459, #3574, #3588); rule `repo.map_gen_stage_protocol` |
 | `tool/check_map_no_partfile_classes.dart` | Two map structure gates share this script. `repo.map_no_partfile_classes` forbids any `part of` file under `packages/colonizethis_map/lib/src/**` from declaring a top-level non-private class. `repo.map_gen_no_new_partfiles` forbids **any** `part` / `part of` directive under the generation layer (`lib/src/gen/**`) and view layer (`lib/src/view/**`) with an **empty** allowlist after #4022 promoted `tile_map_generator_types.dart` to a standalone library (closing the temporary #3588 exception); rules `repo.map_no_partfile_classes`, `repo.map_gen_no_new_partfiles` |
 | `tool/check_map_lib_file_size.dart` | Enforce that every `.dart` file under the colonizethis_map generation layer (`lib/src/gen/**`), view layer (`lib/src/view/**`), and render layer (`lib/src/render/**`) is **550 non-comment lines or fewer** (fail when strictly greater), reusing `countNonCommentLinesFromSource`; the gate also fails when any scanned root is missing so it cannot silently rot if those layers move. Complements the repository-wide `repo.dart_file_non_comment_line_size` cap so the #3588 / #3846 part-file decomposition keeps each pass/concern small (Refs #3588, #3846); rule `repo.map_lib_file_size` |
-| `tool/check_map_lib_file_size_headroom.dart` | Wave-5 headroom ratchet: every `.dart` file under the same gen/view/render scan roots as `repo.map_lib_file_size` must be **350 non-comment lines or fewer** unless listed in the manifest grandfather baseline (shrink-only allowlist). Complements the 500-NCL hard cap without lowering it (Refs #4112, #4297); rule `repo.map_lib_file_size_headroom` |
+| `tool/check_map_lib_file_size_headroom.dart` | Wave-6 headroom ratchet: every `.dart` file under the same gen/view/render scan roots as `repo.map_lib_file_size` must be **300 non-comment lines or fewer** unless listed in the manifest grandfather baseline (shrink-only allowlist). Complements the 500-NCL hard cap without lowering it (Refs #4112, #4297, #4371); rule `repo.map_lib_file_size_headroom` |
 | `tool/check_data_lib_file_size.dart` | Enforce that every hand-written `.dart` file under `packages/colonizethis_data/lib/src/**` is **400 physical lines or fewer** (fail when strictly greater); skips generated suffixes (`.g.dart`, `.freezed.dart`, `.mocks.dart`, `.gen.dart`). Complements the repository-wide 1000 NCL cap so wave-5 topic splits stay small (Refs #4072, #4292); rule `repo.data_lib_file_size` |
 | `tool/check_data_no_part_directives.dart` | Forbid new Dart `part` / `part of` directives under `packages/colonizethis_data/lib/**` outside a shrink-only grandfather allowlist while wave-4 de-part slices land (mirrors app/orders/world no-part gates; Refs #4121) — see `SPEC/program/data-no-part-directives.md`; rule `repo.data_no_part_directives` |
 | `tool/check_data_test_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_data/test/**` must stay at or below **400** physical lines (empty shrink-only allowlist; mirrors `repo.economy_test_file_size`; Refs #4121, #4292) — see `SPEC/program/data-test-file-size.md`; rule `repo.data_test_file_size` |
